Moved and renamed plugin modules. Made plugin more generic as we may add more features later
