+++ plugin/kermit-gradle-plugin/src/main/kotlin/co/touchlab/kermit/gradle/StripSeverity.kt
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2021 Touchlab
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+ */
+
+package co.touchlab.kermit.gradle
+
+enum class StripSeverity {
+    None,
+    Verbose,
+    Debug,
+    Info,
+    Warn,
+    Error,
+    Assert,
+    All
+}
