+++ plugin/kermit-ir-plugin-native/.gitignore
