+++ plugin/kermit-ir-plugin/src/main/kotlin/co/touchlab/kermit/irplugin/KermitComponentRegistrar.kt
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2021 Touchlab
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+ */
+
+package co.touchlab.kermit.irplugin
+
+import com.google.auto.service.AutoService
+import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
+import org.jetbrains.kotlin.cli.common.CLIConfigurationKeys
+import org.jetbrains.kotlin.cli.common.messages.MessageCollector
+import org.jetbrains.kotlin.com.intellij.mock.MockProject
+import org.jetbrains.kotlin.compiler.plugin.ComponentRegistrar
+import org.jetbrains.kotlin.config.CompilerConfiguration
+
+@AutoService(ComponentRegistrar::class)
+class KermitComponentRegistrar(
+  private val defaultStripBelow: String
+) : ComponentRegistrar {
+
+  @Suppress("unused") // Used by service loader
+  constructor() : this(
+    defaultStripBelow = "None"
+  )
+
+  override fun registerProjectComponents(
+    project: MockProject,
+    configuration: CompilerConfiguration
+  ) {
+    val messageCollector = configuration.get(CLIConfigurationKeys.MESSAGE_COLLECTOR_KEY, MessageCollector.NONE)
+    val stripBelow = configuration.get(KermitCommandLineProcessor.ARG_STRIP_BELOW, defaultStripBelow)
+
+    IrGenerationExtension.registerExtension(project,
+      KermitIrGenerationExtension(messageCollector, stripBelow)
+    )
+  }
+}
+
+
