--- stripper/kermit-stripper-plugin/src/main/kotlin/co/touchlab/kermit/stripper/KermitStripperCommandLineProcessor.kt
@@ -1,55 +1,0 @@
-/*
- * Copyright (C) 2020 Brian Norman
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package co.touchlab.kermit.stripper
-
-import co.touchlab.BuildConfig
-import com.google.auto.service.AutoService
-import org.jetbrains.kotlin.compiler.plugin.AbstractCliOption
-import org.jetbrains.kotlin.compiler.plugin.CliOption
-import org.jetbrains.kotlin.compiler.plugin.CommandLineProcessor
-import org.jetbrains.kotlin.config.CompilerConfiguration
-import org.jetbrains.kotlin.config.CompilerConfigurationKey
-
-@AutoService(CommandLineProcessor::class)
-class KermitStripperCommandLineProcessor : CommandLineProcessor {
-  companion object {
-    private const val OPTION_STRIP_BELOW = "stripBelow"
-    val ARG_STRIP_BELOW = CompilerConfigurationKey<String>(OPTION_STRIP_BELOW)
-  }
-
-  override val pluginId: String = BuildConfig.KOTLIN_PLUGIN_ID
-
-  override val pluginOptions: Collection<CliOption> = listOf(
-    CliOption(
-      optionName = OPTION_STRIP_BELOW,
-      valueDescription = "strip log calls below severity",
-      description = "strip severity",
-      required = false,
-    ),
-  )
-
-  override fun processOption(
-    option: AbstractCliOption,
-    value: String,
-    configuration: CompilerConfiguration
-  ) {
-    return when (option.optionName) {
-      OPTION_STRIP_BELOW -> configuration.put(ARG_STRIP_BELOW, value)
-      else -> throw IllegalArgumentException("Unexpected config option ${option.optionName}")
-    }
-  }
-}
